Spring-WebService/src/test/java/com/guifindel/webservice/service/PostsServiceTest.java
package com.guifindel.webservice.service;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

import org.junit.After;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import com.guifindel.webservice.controller.PostsSaveRequestDto;
import com.guifindel.webservice.domain.Posts;
import com.guifindel.webservice.domain.PostsRepository;

@RunWith(SpringRunner.class)
@SpringBootTest
class PostsServiceTest {

	@Autowired
	private PostsService postsService;
	
	@Autowired
	private PostsRepository postsRepository;
	
	@After
	public void cleanup() {
		postsRepository.deleteAll();
	}
	
	@Test
	void DtoData_Save_On_Posts_Table() {
		//given
		PostsSaveRequestDto dto = PostsSaveRequestDto.builder()
				.author("guifindel@gmail.com")
				.content("테스트")
				.title("테스트 타이틀")
				.build();
		
		//when
		postsService.save(dto);
		
		//then
		Posts posts = postsRepository.findAll().get(0);
		assertThat(posts.getAuthor()).isEqualTo(dto.getAuthor());
		assertThat(posts.getContent()).isEqualTo(dto.getContent());
		assertThat(posts.getTitle()).isEqualTo(dto.getTitle());
		
		
				
	}

}
